feat(content): Add sinister assisting mission about a secret slave ship
This commit adds a new assisting mission occasionally offered when
patching up a disabled merchant freighter. In it, the player meets a
sinister fellow named captain Tarquin who is running a secret slave
ship, and is given the option to help him, kill him, apprehend him and
turn him into law enforcement, die, or bungle the situation entirely
and let him escape, in which case he harbors a hatred of the player
that may cause him to pop up later in a future minor mission that is
yet to be written.