Reverse Thrusters
Adds:
- "X1100 Ion Reverse Engines"
- "Capybara Reverse Thruster"

These follow the same reverse thruster pattern as the AR120 atomic, namely they are just the smallest thruster of that series, mounted in reverse. They have been added to a few of the relevant outfitters so they will show up with some of the larger engine sets of their type.