Adding the Mining Laser
This is a short range laser that does a little more damage than a regular laser, but virtually no shield damage and at a 50 range and a slightly smaller size than a regular laser.

The laser beam is simply the regular laser beam scalled to be 100 long instead of 600.

The important detail is that the mining laser has `miner 1`, which triggers the mining behavior, which means if it is installed on a fighter or drone you launch, they will automatically mine.