Creating Light Weapon Pods
Adds:
- Meteor Missile Pod
- Sidewinder Missile Pod
- Javelin Mini Pod
- Heavy Rocket Pod

All of these are roughly 1/4 the size of the main launcher, and hold roughly 1/4 to 1/9 the capacity. All are intentionally worse in missiles-per-space calculations, and utilize the burst fire mechanism to implement a reloading penalty to make it less useful to combine these with racks/boxes.

Graphics courtesy of Becca.

Copyright changes also include two re-orderings to be in alphabetical order.