Upgrading Security Station defense value from 3.4 to 8
Because something that actually costs you 2 outfit space should give you *massively* more defense than something that takes up 0 space.  Pretty sure 8 isn't nearly high enough, but it is a good start.

Also, giving the carrier 5 extra security stations because it has the space and, well, it is a warship. It shouldn't be *easy,* should it?