Misc changes.
Moved the last option of the first choice up one spot, so that the decline option is last. Removed an unnecessary goto on the second choice. Edited the description of the first stadium of the Saryd game. Edited the bit about the announcer in the Saryd game to not be as shallow as it was. Edited the description of how big the seating area of the Kimek stadium is. Reduced the usage of the word "massive."