Adds a mission about bringing a human to Hai space
In Hai space, it is implied that there is a not insignificant amount of travel going on between human and Hai space. We've seen Hai going to human space, so I wanted to write a mission that showed the reverse.
This is my first PR, so the code is almost certainly trash. Any writing/code advice would be greatly appreciated. Thanks to all the lovely people on the Discord that put up with my stupid questions and helped me get this mission to PRable status.