Invert this search
Instead of searching all visited systems to see if they are unshrouded and link to this system, search all links to this system to see if they are unshrouded and have been visited. This reduces the number of systems we need iterate over.

There's some lambda magic going on here that I don't understand but it works.