Tweaked the Emerald Sword's drag and mass
Sheragi engine outfits and the Sheragi fighters are one of a kind, so I decided to just leave them as is and tweak the Emerald Sword slightly to have its default speeds better fit in with where other heavy warships are now at.

If that means that you can create something busted by outfitting the fighters with different engines or using Sheragi engines on other ships then so be it.