Triple all acceleration
Simply tripled all thrust and drag, resulting in max speeds remaining the same but all acceleration being tripled.

The purpose of this is mainly to make fighters accelerate faster. In order to bring down the acceleration of larger ships, should we want to, we'll now need to increase mass (even more) while increasing turn rate by a proportional degree.