Create an "acceleration multiplier" attribute intended for fine tuning of a ship's acceleration
thrust / mass = acceleration, turn / mass = turn rate, and thrust / drag = max speed are too much to juggle. With this, acceleration remains tied to mass because physics, but we can tweak the acceleration if we want without messing with the turn rate or max speed of the ship.

Does it make physical/realistic sense? Not really, but I don't care, I just want balance.