Give toughnesses to minable asteroids
Toughnesses are based on the default average worth of an asteroid, not its hull or the outfit worth. Prospected asteroids have the same ranking as unprospected ones in terms of credit worth.

This doesn't need to be the case, but it's what I've done.

These toughnesses assume a 114 hull dps laser (1.9/frame) with 0.1 prospecting.

The Korath Slicer, instead of creating a separate Korath mining outfit, could have a prospecting of 2.7 for an average 40% boost in drop rates over the human outfit described above, or a prospecting of 3.4 for a 50% boost.