Create unique projectiles for mining lasers
Not really much one can do in the way of laser projectiles.

These are simply beam laser projectiles squashed down to match a range of 200, tinted more toward the red, and with the first frame getting wider as you get farther from the point of firing, giving the effect of the beam fluctuating at it moves between a wide beam and a narrow beam.