Decay the velocity of flotsams along the perpendicular of the pull instead of by some factor based on how perpendicular it is
Behavior should be similar (i.e. the flotsam loses perpendicular motion), the main difference just being the direction of the deceleration.

The previous method measured how perpendicular the pull was from the flotsam velocity and applied a deceleration along the flotsam velocity according to that. Behaviorally this was like applying an additional drag force against the flotsam's direction of travel, with the drag force scaling with how perpendicular the flotsam is to the direction of pull. The constant 0.01 was just an arbitrary scaling factor and its relavence to the outcome was difficult to grasp. For example, if the constant were 1, then the flotsam would still have some perpendicular motion.

This method instead applies a deceleration to the flotsam's velocity that is perpendicular to the direction of pull. Behaviorally this is like pulling the flotsam's velocity toward the pull velocity. The constant 0.01 can be seen as the percentage of how much of the flotsam's perpendicular motion is killed. For example, if the constant were 0.1, then 10% of the flotsam's perpendicular motion is killed. If the constant were 1, then absolutely all perpendicular motion would be killed. I find this difference in how understandable the constant is to be beneficial.