Housekeeping
Restored a tooltip that somehow got deleted.

Reworded the description of when ships become disabled.

The exact equation results in a curve where hull values close to 0 have a disabled threshold of 50% while values close to infinity have a disabled threshold of 10%, but in practice this 50% bound is never actually achieved by any ships in game, so describe a more realistic *about* 45% value for the absolute smallest of ships.

Also describe very large ships as being ships with thresholds close to 10%, as it isn't until you get to the hull strength of a KIV that you actually drop below 11%.