Deleted the old source folder and replaced it with the source from Revamp v4
The problem with updating Revamp to Master was that stuff broke in the engine code.

So, I created a clean branch from fully up-to-date Master, then applied the Formations, Fighter AI, and Parallax changes to it. Once I confirmed that it built & passed the checks, I took that copy of the source folder, and used it to completely replace the source in this original Revamp v1 folder. This is the result of that.