Make use of WeightedList in Fleet for choosing variants
Creates a WeightedVariant class that contains a variant and a weight. WeightedVariant can either have a pointer to a stock variant or a definition of a variant that was defined by the fleet. (This simplifies Fleet to not need to keep track of two separate lists of variants depending on where they came from.) WeightedVariants are held by Fleets in a WeightedList.

The WeightedList now keeps track of the weight total of the fleet instead of the fleet having a separate `total` variable. The WeightedList also takes care of the choosing of a variant.