Improve formation join/leave behavior
Ships now take position in a formation in the order in which they
arrive (instead of the order in the Ships-list). This makes the
formations grow more organically.

Positions of destroyed ships also get back-filled from the end of
the formation. This also reduces the amount of unnecessary intermediate
movement.

The iterators that calculate the ship positions also have been moved
to the Step() function. This could allow the actual position-retrieve
functions to become const in the future.