fix(content): Fix civilian and pirate ship phrase weightings
In the port to use phrase weightings, phrases were weighted on a 1:1
basis with how many names are possible in that phrase. This results in
phrases with many options being hugely over-weighted, with less-common
phrases being rarely seen. The result is that civilian and pirate ship
names follow the same patterns too frequently. This commit fixes that by
reducing the frequencies of the phrases with the most options. This
approximately returns the phrase weightings to about whet they were
before adopting manually-weighted phrasings.