Hai stations review from Amazinite (#30)
* Fix copyright entry

* Simplify news location filters

No need to put the `not` and `attributes` on separate lines.

* Don't offer missions on uninhabited stations

* Reorder planet attributes

Super nitpicky. Regional attribute goes first.

* Fix up Heia Due

The new stations in Heia Due caused sprite overlap. Moved the first planet closer to the system center so that it doesn't clip with the second station and removed Coppermist because it was so close to the system center that it would clip into the stars. Opted to remove instead of move elsewhere because it was a mothballed station, and we seem to have plenty of those anyway.

* Overlap avoidance in Fah Soom

* Overlap avoidance in Wah Oh + environmental storytelling

Changed the period of Ochrescoop to match the period of the object it orbits, causing the station to always be on the opposite side of the planet relative to the planet's orientation to the star. Wah Oh is an extremely hot star with a high habitability range, so having a station that close to the star could cause problems. Having it be permanently in the shadow of the planet it orbits would help alleviate any such issues.

The periods being as close as they were already is what gave me this idea.

* Overlap avoidance in Io Mann

Nuked Lightstead and moved its description over to Thornlight.

* Overlap avoidance in Due Yoot

Nuked Indigosprint and moved Greenview to where Indigosprint was.
Also nuked an uninhabited planet that could ever so slightly clip with Greenview.

* Overlap avoidance in Rati Cal

Moved Limejob to orbiting the outermost planet, as the inner system was pretty crowded.

* Overlap avoidance in Da Lest

* Overlap avoidance in Lom Tahr

* Overlap avoidance in Io Lowe

* Overlap avoidance in Ya Hai

* Overlap avoidance in Da Ent

* Overlap avoidance in Uwa Fahn

* Overlap avoidance in Ula Mon

* Nuked Icemeet

Another station that clipped the star. This system seems crowded enough already so I'm not even going to bother moving the station somewhere else in the system.

* Overlap avoidance in Imo Dep

* Overlap avoidance in Bote Asu

* Overlap avoidance in Zuba Zub

* Nuked Silverleap

* Description updates

Patched up some grammatical errors and fixed a reference to a station's location that was no longer accurate.

Word of the day: promenade.