Move Test::Context to its own TestContext class.
If the test-context is in its own class, then it is possible to refer to
such class with a forward declaration. This is beneficial if the context
needs to be transferred through other classes that only need to forward
pointers or references, since such classes can then use forward declarations
to make the class internally known.