Transport Missions -> Generic Missions :P
From you earlier: [...] the point of the name is to describe the nature of the contents.
And later: We can go further in describing them as conversations that occur in the spaceport for the purpose of giving the player more insight into the culture of the planet they've landed on that don't lead to any mission or objective, but contained within that is the description of a spaceport conversation.

I contest that they're actually conversations for the purpose of giving the player more insight into the culture of the area they're in, that typically occur in the spaceport.

I would rather put the in-space convos and other things I have up my sleeve here, and let non-cultural conversations go to generic missions, than the reverse. Thematically the purpose of the file is definitely stronger with culture conversations.