Add notes on adding operator[] to ConditionsStore
We are getting near the point were we could support the operator[] on
ConditionsStore for Value, Exact and Prefix ConditionEntries.

A reference to the ConditionEntry would then be returned from this
operator and the functions mentioned here would be used to automatically
convert the ConditionEntry from and towards int64_t values.