fix(integration-test): Remove Panel dependencies from Test.cpp / Test.h (#6295)
* Remove Panel dependencies from Test.cpp / Test.h

The tests don't need to actually see the panels, and this change allows for moving the Test call to the thread-safe section.

Replace menuPanels.Quit() by SDL_QUIT event in Test.cpp.
Place command-forwarding outside Test.cpp.

* Update comments to reflect the removal of Panels.

* Move Test::Context to its own TestContext class.

If the test-context is in its own class, then it is possible to refer to such class with a forward declaration. This is beneficial if the context needs to be transferred through other classes that only need to forward pointers or references, since such classes can then use forward declarations to make the class internally known.

* Add TestContext.h for XCode compilation.

* Thread-safe test-control during flight.

Allow main.cpp to transfer test-control to the thread-safe part of Engine during flight to allow test-control to be executed in a thread-safe manner while the Engine is running multi-threaded.

* Apply suggestions from code review

Co-authored-by: Ben Hauch <tehhowch@users.noreply.github.com>

* Process review comments

- Changed a number of includes to forward declarations.
- Have test-code add commands instead of overwriting them.
- Improved some comments.
- Remove unused parameter.
- Change some variables from protected to private.

* Provide some helper functions in TestContext.

Based on review comments.

* Update XCode project to add TestContext.cpp

* Use default constructor as suggested during review.

* Initialize vector directly in constructor.

Co-authored-by: Ben Hauch <tehhowch@users.noreply.github.com>