Mining ships choose a random belt to search
Belts are chosen to be searched at the same probability as asteroids are likely to be in that belt.

Without this, mining ships would search at the average distance of all belts, which could be outside of the range of any minables, resulting in them never finding an asteroid to mine.