Planetary Tweaks
Noting this assumes that savanna specifically is the preferred climate. IMPORTANT: A savanna is defined as a grassland dominated *tropical or subtropical* region which experiences monsoonal type weather. I.e. it's hot and dry part of the year, and hot and very wet the other part.

I.e. Kimek like it hot and dry, as long as they have access to water (perhaps they burrowed for it during dry seasons in their prehistory), and will mostly complain if it rains consistently throughout the year, rather than if it rains at all.