Nerfing the scan penalty
I agree with the comments that the scan penalty was a bit high. That being said, I still think there needs to be a penalty, not just provoke. Just provoke is rather toothless, rather like the post-wanderer non-hostile automata that one can attack, capture, and plunder at will. Provoking there is meaningless.