Recompute the habitable range of all systems given the stars they have
Also tweaked the types of stars and distances of planets in some systems, mainly in inhabited systems in order to keep planets about where they were before in terms of habitability.

The habitable range of a system influences the period of the objects in that system, hence the period changes.

There are still a handful of systems that need addressing, but this commit covers 98% of systems.

Notable changes:
* All ringworld systems now use g3 stars instead of g5-old.
* Made significant changes to Polaris to better match the real world star type.

Systems to still address:
* Suhail
* Betelgeuse
* Alpha Centauri
* Almaaz
* Mebsuta
* Naos
* Algol
* Alnitak

The above systems will have the proper habitable range and orbital period, but the planets in those systems are likely scalding hot. Need to address the star types used and/or planet distances in those systems.