Organic Arrival Distances
Alright, the problems with the script were resolved, so now the arrival distances get added in the right place.

So, the script finds "habitable distance + 150" and "belt distance + 150" and compares the two. Whichever distance is greater it assigns as the arrival distance.

Most systems have an arrival distance between 1500 and 2000, but there's a decent number of outliers on both sides of that.