Tweak some numbers
o-type stars now use radii of 70 to 80 instead of 70 to 85.

Carbon stars now use a habitable of 625.

Neutron stars (unused) now use a habitable of 100.

l-dwarves now use a habitable of 30 instead of 45; the previous value was larger than the m-dwarf because of the image sizes, but l-dwarves are actually cooler than m-dwarves.

The system editor uses a defined radius for each star to determine its habitable range and its mass, which is what is used for determining the period of objects in the system. Some stellar objects, such as nova remnants or neutron stars, should have masses higher than their habitable ranges, though. As such, I've divorced habitable range and mass for novas, neutron stars, and black holes. The former three now use a mass equivalent to an a0 star while the black hole uses a mass double that.