Separate position and direction into separate methods.
The position generation can be done once every 20 steps,
or whenever an event happens that requires recalculation.

The direction calculation should still be done every step
to ensure quick responses.

Flipping now also moved to the position calculation, since
it doesn't need to be calculated every step (but only when a
flip happens).

Tallying now also moved to the position calculation, since
it doesn't need to be done every time a position is retrieved.