Better safe than sorry
In order to ensure that Endless Sky's data structure doesn't risk breaking (or being broken by) things that don't understand folders, file names, or directories that include non alpha-numeric characters, removing the apostrophe from the ka'het name in the directory and file names only.
This does NOT change the data or content, only the names of the folder and files.