Editing extreme values
As Derpy pointed out, a few star systems have extreme habitable and/or belt ranges. (as in, over 10k). This editing pass is me going through and finding anything that looks particularly big (pretty much anything over 4k, and selectively changing them. In most cases, to 150 over whatever the largest orbit is in the system.