Providing telemetry explanation
For this rescue mission, the player receives telemetry info in the form of the target ship being listed as an escort. The description, however, makes no mention of this, so it looks like a bug.

This tweaks the description so that it mentions that there is telemetry data to follow.