Share a single DamageProfile across all ship impacted by an explosion or hazard
DamageProfile no longer receives a specific ship in its constructor, only receiving information about the weapon that is dealing damage.

The DamageProfile constructor precomputes any information that is independent of the ship being impacted. This includes precomputing blast radius variables if the weapon is an explosion and damage dropoff scaling if the weapon is not a hazard and has a damage falloff.

CalculateDamage then receives a reference to the ship being damaged and finishes computing the damage scale; blast radius calculations are finished using the ship's information and damage dropoff calculations are run if they were deferred during construction. (This occurs when a DamageProfile is created by a hazard, as the distanceTraveled variable is ship-dependent for hazards.)

Finally, damage is calculated for the specific ship given. On each CalculateDamage call, the values for the previous ship get written over, so ships don't influence each other in terms of damage received.