Tech Debt Clean-Up
This wraps up the remaining "tech debt" in the Remnant missions so that they are all consistent. As a benefit, this also means they will sort properly together, be easier to find, easier to read in save-files, and easier to use as they conform to the standard Remnant mission & event structure.