fix(internal): Issue indentation warnings correctly (#6422)
* Identify the correct line where mixed whitespace is used in a file

This includes warning about mixed whitespace used in lines that are only comments. Such lines are only identified by their line number instead of giving a typical PrintTrace.

Also warn about any switch in indentation in a file instead of only warning about switching from spaces to tabs. This also removed the "only warn once per file" behavior that only occurred under certain circumstances.
(In a tabs file, "tab space" would be warned about every time but only the first line starting in a space would be warned about. Not only did this seems off to me, but only warning about the first line also meant that you'd have to relaunch the game to get the warning about the next line if a file for some reason had multiple such issues scattered throughout the file.)

Also moved the empty line for PrintTrace to the end instead of the beginning because it irked me.

* Fix whitespace and standardize print traces

* Add unit tests for DataFile parsing (amazinite/endless-sky#22)

* Use inclusive language

Co-authored-by: Nick <quyykk@protonmail.com>
Co-authored-by: Ben Hauch <tehhowch@users.noreply.github.com>