Nested variants have different validity requirements
A top-level variant is valid if when chosen it could spawn a single valid ship. This means that either one of its ships is valid or one of its nested variants will produce a valid ship. But a nested variant is only guaranteed to produce a valid ship if all of its ships or nested variants are valid because of how one is chosen at random.