Minimum necessary change + going the full mile (#117)
* Minimum necessary change

We've never had the standard of ensuring compatibility with plugins, so don't bother with the extra stuff that only applies to plugins.

In changing the event name for changing hails, we can actually cut out an older compatibility mission.

In renaming the "Remnant defense delay" condition, I've also made it all lowercase to match other set conditions.

Going the full mile and patching some parts that were missed. Changes in the Cognizance file require no compatibility handling as they've never been released.

* Missed some

* Update data/remnant/remnant events.txt

Co-authored-by: Zitchas <32169904+Zitchas@users.noreply.github.com>

* Allow some one-time duplicates of non-conversational missions

We most definitely want to avoid reoffering a mission with a conversation, but these two aren't those such missions, so it's less important that they be blocked from offering if the older version of the mission offered. This cuts down on the amount of compatibility code that needs to be maintained.

* Even lower minimum

* Even LOWER minimum

* Ditto

Co-authored-by: Zitchas <32169904+Zitchas@users.noreply.github.com>