Housekeeping
Improve comment under the on aborts.

Rearranged offer conditional

Added a missing mission name and description to one of the spaceport missions.

Adding a missing decline at the end of a spaceport on offer conversation.

Reformatted all the missions in the file. Yes I know a bunch of missions in the game don't follow this format. No I'm not going to finish this any time soon.