Remove first mission's reliance on Dusk
As mentioned in my original review back in October, blocking this line of missions entirely on Dusk not being busy seems like it could be jarring. This change makes it so that the first mission doesn't rely on Dusk being available, moving that requirement to the second mission. The second mission will then react to whether Dusk was busy during the first mission.

Part of this change has also involved chopping up the on enter conversation from the first mission and moving some of it into the on complete as to have the reading be less dense.