Adjust second mission to introduce Dusk into the mix
This takes the old on complete conversation from the first mission and makes it the on offer conversation of the second mission while also accounting for whether Dusk was previously busy.

This commit also changed the first mission from a landing to a minor mission that asks you if you want to find Taely, as opposed to going right to her.