feat: Disabled damage type (#4404)
Adds disabled damage type.

Disabled damage is the amount of damage a weapon does to a ship that is disabled. Disabled damage is proportional to how much damage left in the weapon on that shot. Ex. if 60% of the damage killed the rest of the shields, and 20% of the damage lowered the hull enough to be disabled, then the final 20% of damage is then applied to the hull and multiplied by the disabled damage attribute.

Co-authored-by: Amazinite <jsteck2000@gmail.com>
Co-authored-by: Peter van der Meer <peter.vd.meer@gmail.com>