Technical correction cement to concrete
Cement is one chemical used to make concrete.
Concrete is the end-product that we routinely see in the form of walls, floors, parking pads, etc.
Cement itself is almost never encountered on its own, so in common culture people commonly conflate cement with concrete, not realizing that the former is just one ingredient in making the later, the two are not actually synonymous.