Mining and fleet AI updates
Intent
------

A note from @samrocketman (Sam Gleske).

This change only introduces engine changes to enable plugins to implement
the behavior initially.  However, eventually the game data files would
also be updated to support the behavior.

All data file changes have been stripped from this code in order to limit
how much gameplay is affected.  You can include the data file changes by
installing the following plugin which includes behavior originally
introduced with this engine change.

The idea is to polish this a bit and to match the code style with the
rest of the project so that it may be merge-worthy.

Primary changelog
-----------------

* Adding Strafing & Circling AI support
  > Thanks to VitalChip for coding this.  They say they are going to redo
  > it better in a few weeks. But in the mean time, this is what I pulled
  > from their branch. Hopefully it works!

Fixes provided by contributors

* Removing the python script
* Removing duplicate arrival
* Missing }
* Fixing duplication of keys
* Fixing a bracket
* And a missing else
* Fixing another bracket

Sources of changes
------------------

* https://github.com/endless-sky/endless-sky/pull/5772
* https://github.com/Zitchas/endless-sky/pull/67
* https://discord.com/channels/251118043411775489/431496424992014347/953417198821580850

Co-authored-by: Zitchas <32169904+Zitchas@users.noreply.github.com>
Co-authored-by: arkhne <36457239+arkhne@users.noreply.github.com>
Co-authored-by: Sam Gleske <sam.mxracer@gmail.com>