Bugfix: restore purchase into cargo
Introduction of bugfix
----------------------

This restores a previously working behavior.  With the introduction of PR #6059 purchasing outfits into cargo or moving outfits from storage into cargo is no longer possible.  This (even if temporarily) restores the old behavior for drawing the `Cargo` button.

Reasoning for this fix
----------------------

The old code shows setting and unsetting `playerShip`.

https://github.com/endless-sky/endless-sky/blob/6348b5ba9d08a2e5f671d133cf7f1138b8267003/source/OutfitterPanel.cpp#L790-L812

The decision to render `Cargo`, `Install`, or `Buy` depends on the result of `IsAlreadyOwned()` and `playerShip`.

https://github.com/endless-sky/endless-sky/blob/949760972eb3053173f2c89001b4ab936e29c500/source/ShopPanel.cpp#L356

See also `IsAlreadyOwned()` source.

https://github.com/endless-sky/endless-sky/blob/949760972eb3053173f2c89001b4ab936e29c500/source/ShopPanel.cpp#L538-L548

Related issues
--------------

See also #6717 (I filed a dupe) and #6668.

Testing done
------------

I was able to do the following with this change:

- [x] Toggle stored outfits, cargo outfits, and outfits for sale independently.
- [x] Purchase outfits into cargo successfully.
- [x] Move outfits from storage into cargo successfully.