Velocity instead of speed; acceleration projectiles expanded
Previously the information saved through submunition layers was stored as speed to be used on the current heading when submunitions are launched.  Now the headings are baked into the saved speed, leading to saved velocity (how it works).  Also, projectiles with acceleration are not branched on anymore but included with regularly fired submunitions.