Support the Home and End keys in the panels
This supports going to the beginning and end of the lists. This is
useful when there are a lot of entries (e.g. items or ships) and you
want to quickly go to something specific, where you know where it is
located (e.g. a newly captured ship in the PlayerInfoPanel or the
attributes of your flagship in the outfitter panel.

The implementation needs to decide whether there should only be
scrolling or whether the first or last element should also be selected.
The used criteria is:
Does something else change when the selection is changed?
Therefore, only the BoardingPanel selects, which is consistent to how
other line-based GUI elements typically behave (e.g. in a file explorer
or a text file). The implementation of the page up and page down keys in
the BoardingPanel have been adapted to reflect this behavior for
consistency reasons.