Hardpoint Offset Oversight
As Amazinite pointed out. (Should've gotten onto this about three days ago!)
This properly accommodates fire effects on hardpoints with offsets. This accommodates both y-offset-only, and x&y-offset turrets. The closest outfit you'd probably see this on in vanilla is the beam turret which has the lasers mounted on one side.