Clean up code and introduce stricter checks
Added helper methods for inaccuracy which can easily be expanded to handle an IsUniform toggle; the code is already there it just needs to be uncommented to accept the variable (which does need to be added).

Checked if the weapon was a spinup type before doing the extra math.  It isn't that intensive so it may be an extra comparison for nothing but most weapons probably won't be spinup so it's probably a good idea anyway.