Remove unecessary spinup tag
Removed spinup tag as suggested.  Specifying without a tag is doable but not preferable, as most (I think all?) key values are locked at a minumum of 1 making simple comparison impossible.

For example, burstCount is locked at 1 to avoid errors in burst logic and spinupTime is locked at 1 to avoid division by 0.  So, it ends up looking like if(SpinupTime() > 1) instead of if(IsSpinup).