Narrow scope and remove ApplyInaccuracy functions
As pointed out, the entire Hardpoint scope is not necessary in order to get a single value.  That value is simply passed through instead.

Also, the ApplyInaccuracy methods no longer implement redundant behavior due to the addition of spinup, and, in the current state, can't even be used in the submunition-specific constructor.