Add stat  to lock burst ratios
It was previously impossible to exactly and intuitively maintain the bursting time to cooldown time ratio on a burst weapon with spinup.  It could have been approximated with some effort and found precisely if you performed the sums, but that would also assume a full burst was fired without interruption.

When present on a burst weapon, the new reload ratio stat supersedes the reload stat.  It represents the multiple of the increment for the burstReload counter to use for the reload counter.