Fix broken directory detection for android assets.
Android Asset api doesn't behave as documented, at least on my device.
Add some additional checks to guess whether or not the directory is
valid.

This was breaking new player creation, since it was trying to detect if
the player existed, but the folder api kept saying yes, and so it would
append a number to the player, and try again, locking up the app.