Add explicit exit() when main is done
Since on android main() is just a function called from SDLActivity after
libmain has been loaded, it is possible for it to be called a second
time if the program is restarted. We could theoretically take advantage
of that if the code was reentrant, but we have such heavy reliance on
static and global variables that it just causes problems or even
explicitly crashes.

For now, just explicitly exit so that if the app is restarted, it will
re-initialize all of the static variables.