Make negative smoothness useful
Negative smoothness will cause projectiles to cluster toward the minimum and maximum inaccuracies.

A coefficient will still be generated using the normal distribution, but the halves of the coefficient will be swapped.  That is, the uncommon 0.1 will become the common 0.6 and the common 0.6 will become the uncommon 0.1.  Now, after this translation the probabilities are flipped, resulting in opposite spread.